requirement.txt
apturl==0.5.2
asgiref==3.5.2
astroid==2.12.13
autopep8==2.0.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
comtypes==1.1.14
contourpy==1.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
defer==1.0.6
dill==0.3.6
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.7
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
duplicity==0.8.12.0
easygui==0.98.1
entrypoints==0.3
fasteners==0.14.1
filelock==3.9.0
filetype==1.2.0
fonttools==4.38.0
future==0.18.2
gcloud==0.17.0
googleapis-common-protos==1.58.0
httplib2==0.14.0
idna==2.8
imageio==2.23.0
isort==5.11.4
jws==0.1.3
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.8.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.7.0
monotonic==1.5
netifaces==0.10.4
numpy==1.24.1
oauth2client==3.0.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
paramiko==2.6.0
pexpect==4.6.0
Pillow==9.3.0
platformdirs==2.6.0
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.10.0
pycryptodome==3.4.3
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.15.9
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
Pyrebase==3.0.27
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-jwt==2.0.1
pyttsx3==2.90
pytz==2022.6
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.11.1
requests-toolbelt==0.7.0
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.3
systemd-python==234
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.19.0
virtualenvwrapper-win==1.2.7
wadllib==1.3.3
wikipedia==1.4.0
wrapt==1.14.1
xkit==0.0.0
